Adding useGameAction and useVocabFuzzyMatcher. Removing GameUtil. Updating the matching game components.
